[IMP] web: merge quick_add with quick_create_form_view id in views
https://github.com/odoo/odoo/pull/114827 added a new attribute
quick_create_form_view_id to the calendar view that would allow showing a form
view when clicking on quick_create. This attribute was only used in calendar,
however it could be useful for other modules as well. It was decided to merge
this new attribute with the quick_add attribute, as they are related. Now, the
quick_add attribute accepts either a boolean or a form view id. This commit
changes the documentation of this attribute.

task-3286969